Update modules/Import/ImportDuplicateCheck.php
An index can use a custom dupeCheckFunction during an import. The problem is that a user can select that index along with other indices, but as soon as the dupeCheckFunction is called it simply returns even if the function does not find a duplicate (returns false). This then bypasses the other additional fields that are wished to be used to check for duplicates. 

Bug # coming shortly.