Update EditView.tpl
If the value is blank/null in the database and a record is edited the UI will show that the default value(s) are selected. This is not desired especially when there is a blank option for an enum or when a user deselects all options on a multienum. It creates what is an unexpected behavior for the user.